+++ app/renderer/suite.js
@@ -13,9 +13,28 @@ class Suite {
     removeTask(index) {
         this.tasks.splice(index, 1);
     }
-    
-    replaceTask(index,task){
-        this.tasks.splice(index, 1,task);
+
+    replaceTask(index, task) {
+        this.tasks.splice(index, 1, task);
+    }
+
+    stopAll() {
+        let promises = [];
+        for (const task of this.tasks) {
+            if (task.isRunning()) {
+                promises.push(new Promise((resolve) => {
+                    task.stop(() => {
+                        resolve();
+                    });
+                }));
+            }
+        }
+        return Promise.all(promises);
+    }
+    runAll() {
+        for (const task of this.tasks) {
+            if (!task.isRunning()) task.run();
+        }
     }
 
     toJSON() {