+++ main.js
@@ -2,7 +2,8 @@
 
 const {
     app,
-    BrowserWindow
+    BrowserWindow,
+    ipcMain
 } = require('electron');
 const path = require('path');
 
@@ -13,14 +14,18 @@ function isDevEnv() {
     return process.env.NODE_ENV === "dev";
 }
 
+ipcMain.on('close-app', () => {
+    app.quit();
+});
+
 //Global reference to window
 let win;
-
 
 function createWindow() {
     const iconPath = path.join(__dirname, 'resources', 'icon.png');
 
     UserConfig.loadConfig((config) => {
+        if (!config.windowSize) config.windowSize = [500, 600]; //TODO: use default
         let winConfig = {
             width: config.windowSize[0],
             height: config.windowSize[1],
@@ -42,7 +47,15 @@ function createWindow() {
         if (isDevEnv()) win.webContents.openDevTools();
 
         win.on('resize', () => {
-            UserConfig.config.windowSize = win.getSize();
+            let size = win.getSize();
+            if (isDevEnv()) size[0] -= config.devToolsSize;
+            UserConfig.config.windowSize = size;
+        });
+        let first = true;
+        win.on('close', (ev) => {
+            win.webContents.send('before-close');
+            if (first) ev.preventDefault();
+            first = false;
         });
         win.on('closed', () => {
             win = null;